:recycle: (Detail): Timetable - @ViewBuilder -> struct
