+++ whenMyBusGo/Sources/Presentation/BusDetailSheet/BusDetailSheetView.swift
@@ -16,8 +16,7 @@ struct BusDetailSheetView: View {
                 busInfo()
                     .frame(maxHeight: 100)
                 
-                timetable(busData.upboundTimetable[.sat])
-                    .background(.white, ignoresSafeAreaEdges: [])
+                Timetable(timetable: $busData.upboundTimetable[.sat])
                 
                 // footer
                 ZStack {
@@ -70,58 +69,6 @@ struct BusDetailSheetView: View {
             .padding()
             .foregroundStyle(.white)
         }
-    }
-    
-    @ViewBuilder
-    private func timetable(_ timetable: [String]?) -> some View {
-        let morningTimes = timetable?.filter { time in
-            guard let hour = Int(time.prefix(2)) else { return false }
-            
-            return hour < 12
-        } ?? ["시간표가 비어있습니다."]
-
-        let afternoonTimes = timetable?.filter { time in
-            guard let hour = Int(time.prefix(2)) else { return false }
-            
-            return hour >= 12
-        } ?? ["시간표가 비어있습니다."]
-
-        let columns = [
-            GridItem(.flexible()),
-            GridItem(.flexible()),
-            GridItem(.flexible())
-        ]
-        
-        ScrollView {
-            VStack(alignment: .leading) {
-                Text("오전")
-                    .font(.title)
-                    .fontWeight(.semibold)
-                
-                LazyVGrid(columns: columns, spacing: 15) {
-                    ForEach(morningTimes, id: \.self) { time in
-                            Text(time)
-                                .font(.title2)
-                    }
-                }
-                
-                Divider()
-                    .padding([.top, .bottom])
-                
-                Text("오후")
-                    .font(.title)
-                    .fontWeight(.semibold)
-
-                LazyVGrid(columns: columns, spacing: 10) {
-                    ForEach(afternoonTimes, id: \.self) { time in
-                            Text(time)
-                                .font(.title2)
-                    }
-                }
-            }
-            .padding()
-        }
-        .background(.white, ignoresSafeAreaEdges: [])
     }
 }
